refactor(config): point the persistence/reload dependency one way
config_persistence imported ReloadContext and ReloadOutcome from
config_reload, while config_reload called back into config_persistence
for write_config and load_config_or_default. Neither module could be
read on its own, and neither could be changed without opening both.

The reload-aware wrappers — write_config_and_reload,
write_config_and_reload_outcome, ConfigMutation, mutate_config_and_reload
— move to config_reload, which already depended on config_persistence.
config_persistence is now pure file I/O with no reload import at all.
No behaviour changes; the bodies are unchanged apart from the paths.

Public paths change for those four items
(config_persistence::* -> config_reload::*). Deliberately no re-export
shim left behind: a `pub use` pointing back at config_reload is the
cycle again wearing a different hat. load_config_or_default and
write_config stay put, so the re-export in gateway/ui/backend_ops.rs is
untouched. One caller needed updating, gateway/ui/backends.rs.

RED before the fix, from the new guard test:
  config_persistence reaches back into config_reload, restoring the
  cycle: ["use crate::config_reload::{ReloadContext, ReloadOutcome};"]

That guard exists because nothing in the build complains about a
module cycle in Rust — only a test can hold this boundary.
