+++ src/config_persistence.rs
@@ -7,7 +7,6 @@ use std::path::{Path, PathBuf};
 use std::sync::atomic::{AtomicU64, Ordering};
 
 use crate::config::Config;
-use crate::config_reload::{ReloadContext, ReloadOutcome};
 
 /// Load config from `path`, returning `Config::default()` when the file is absent
 /// or cannot be parsed.
@@ -48,94 +47,6 @@ pub fn write_config(path: &Path, config: &Config) -> Result<(), String> {
     let yaml =
         serde_yaml::to_string(config).map_err(|e| format!("Failed to serialize config: {e}"))?;
     write_yaml(path, &yaml)
-}
-
-/// Serialize `config`, write it atomically, then trigger hot-reload when a
-/// reload context is available.
-///
-/// Persistence is always authoritative for the on-disk file. Hot-reload then
-/// applies only the subset of changes supported by [`ReloadContext`] (for
-/// example, backend changes); server listener changes remain on disk until the
-/// process is restarted.
-///
-/// # Errors
-///
-/// Returns an error string on serialization, write, rename, or reload failure.
-pub async fn write_config_and_reload(
-    path: &Path,
-    config: &Config,
-    reload_context: Option<&ReloadContext>,
-) -> Result<(), String> {
-    write_config_and_reload_outcome(path, config, reload_context)
-        .await
-        .map(|_| ())
-}
-
-/// Serialize `config`, write it atomically, then return any hot-reload outcome.
-///
-/// # Errors
-///
-/// Returns an error string on serialization, write, rename, or reload failure.
-pub async fn write_config_and_reload_outcome(
-    path: &Path,
-    config: &Config,
-    reload_context: Option<&ReloadContext>,
-) -> Result<Option<ReloadOutcome>, String> {
-    if let Some(ctx) = reload_context {
-        // Write and reload share one lock inside the context. Writing here
-        // first would reopen the race the lock exists to close.
-        return ctx.write_and_reload_outcome(path, config).await.map(Some);
-    }
-
-    write_config(path, config)?;
-    Ok(None)
-}
-
-/// What a guarded read-modify-write did: either the change was applied and
-/// persisted, or the caller's own check rejected it and nothing was written.
-pub enum ConfigMutation<T, E> {
-    /// The change was applied, persisted, and (when a reload context exists)
-    /// reloaded.
-    Applied(T, Option<ReloadOutcome>),
-    /// The caller's closure refused the change. The file is untouched.
-    Rejected(E),
-}
-
-/// Read the config, apply `mutate` to it, and persist the result without
-/// letting another writer slip in between the read and the write.
-///
-/// Reading outside the lock is what makes edits vanish: two requests each read
-/// the same starting file, each apply their own change to that stale copy, and
-/// the second write erases the first change while reporting success. Doing the
-/// read inside the same critical section as the write is what stops it.
-///
-/// # Errors
-///
-/// Returns an error string on validation, write, rename, or reload failure. A
-/// refusal from `mutate` is not an error; it comes back as
-/// [`ConfigMutation::Rejected`] with the file untouched.
-pub async fn mutate_config_and_reload<T, E, F>(
-    path: &Path,
-    reload_context: Option<&ReloadContext>,
-    mutate: F,
-) -> Result<ConfigMutation<T, E>, String>
-where
-    F: FnOnce(&mut Config) -> Result<T, E>,
-{
-    if let Some(ctx) = reload_context {
-        return ctx.mutate_and_reload_outcome(path, mutate).await;
-    }
-
-    // No live gateway to reload, so no reload lock exists to hold. This path is
-    // the CLI acting on a config file nothing else is serving.
-    let mut config = load_config_or_default(path);
-    match mutate(&mut config) {
-        Ok(value) => {
-            write_config(path, &config)?;
-            Ok(ConfigMutation::Applied(value, None))
-        }
-        Err(rejection) => Ok(ConfigMutation::Rejected(rejection)),
-    }
 }
 
 /// How many times a rename is retried before the write is reported failed.
@@ -482,31 +393,5 @@ mod tests {
 
         assert!(matches!(result, Err(msg) if msg.contains("Failed to validate config")));
         assert!(!path.exists());
-    }
-
-    #[tokio::test]
-    async fn write_config_and_reload_without_context_persists_yaml() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("gateway.yaml");
-        let config = Config::default();
-
-        write_config_and_reload(&path, &config, None).await.unwrap();
-
-        assert!(path.exists());
-        let loaded = Config::load(Some(&path)).unwrap();
-        assert_eq!(loaded.backends.len(), config.backends.len());
-    }
-
-    #[tokio::test]
-    async fn write_config_and_reload_outcome_without_context_returns_none() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("gateway.yaml");
-        let config = Config::default();
-
-        let outcome = write_config_and_reload_outcome(&path, &config, None)
-            .await
-            .unwrap();
-
-        assert!(outcome.is_none());
     }
 }
